#127: added exception throwing
